restructured files, created Content comp
