+++ src/components/App/App.scss
@@ -1,5 +1,6 @@
 // @import "~bootstrap/scss/bootstrap";
 
-// .App {
-//   text-align: center;
-// }
+.App {
+  background-color: #F8F8F8;
+  padding-top: 30px;
+}
