+++ src/components/Content/Content.scss
@@ -1,0 +1,4 @@
+.content {
+  margin-top: 50px;
+  background-color: white;
+}
